requirements.txt
neo4j-driver
requests
slackclient